refactor(storage): honest names for hydration and board-set delete
Rename queries.ts -> board-hydration.ts: the file's reason to exist is
OBF->Board hydration, not querying. Drop two zero-behavior passthroughs
(getBoardSet, persistBoardTranslations) — callers now use db directly,
which is already the de-facto persistence API.

Rename db.deleteBoardSet -> deleteBoardSetRecord to end the name clash
with the store's delete+notify wrapper. Removes the alias that papered
over the collision at the import site; the two now read as the record
primitive vs. the observable mutation they are.
